requirements.txt

pyyaml==6.0
suggestion==0.0.6
n
